requirements.txt
streamlit>=1.40
supabase>=2.10
bcrypt>=4.2
pandas>=2.2
python-dotenv>=1.0
pytz>=2024.1
